move is_citable
