+++ src/almanack/metrics/garden_lattice/connectedness.py
@@ -12,7 +12,7 @@ import pygit2
 import requests
 import yaml
 
-from ...git import find_file, read_file
+from ...git import file_exists_in_repo, find_file, read_file
 from ..remote import get_api_data
 
 LOGGER = logging.getLogger(__name__)
@@ -118,6 +118,59 @@ def detect_social_media_links(content: str) -> Dict[str, List[str]]:
         "social_media_platforms": sorted(found_platforms),
         "social_media_platforms_count": len(found_platforms),
     }
+
+
+def is_citable(repo: pygit2.Repository) -> bool:
+    """
+    Check if the given repository is citable.
+
+    A repository is considered citable if it contains a CITATION.cff or CITATION.bib
+    file, or if the README.md file contains a citation section indicated by "## Citation"
+    or "## Citing".
+
+    Args:
+        repo (pygit2.Repository): The repository to check for citation files.
+
+    Returns:
+        bool: True if the repository is citable, False otherwise.
+    """
+
+    # Check for a CITATION.cff or CITATION.bib file
+    if file_exists_in_repo(
+        repo=repo,
+        expected_file_name="citation",
+        check_extension=True,
+        extensions=[".cff", ".bib"],
+    ):
+        return True
+
+    # Look for a README.md file and read its content
+    if (
+        file_content := read_file(
+            repo=repo, filepath="readme.md", case_insensitive=True
+        )
+    ) is not None:
+        # Check for an H2 heading indicating a citation section
+        if any(
+            check_string in file_content
+            for check_string in [
+                # markdown sub-headers
+                "## Citation",
+                "## Citing",
+                "## Cite",
+                "## How to cite",
+                # RST sub-headers
+                "Citation\n--------",
+                "Citing\n------",
+                "Cite\n----",
+                "How to cite\n-----------",
+                # DOI shield
+                "[![DOI](https://img.shields.io/badge/DOI",
+            ]
+        ):
+            return True
+
+    return False
 
 
 def find_doi_citation_data(repo: pygit2.Repository) -> Dict[str, Any]:
