+++ src/book/garden-lattice/understanding.md
@@ -1,1 +1,2 @@
 # Understanding
+
