+++ tests/metrics/test_data.py
@@ -18,18 +18,22 @@ from almanack.metrics.data import (
     METRICS_TABLE,
     _get_almanack_version,
     compute_repo_data,
-    count_repo_tags,
+    get_api_data,
+    get_github_build_metrics,
+    get_table,
+    measure_coverage,
+)
+from almanack.metrics.garden_lattice.connectedness import (
     count_unique_contributors,
     default_branch_is_not_master,
     find_doi_citation_data,
-    get_api_data,
-    get_ecosystems_package_metrics,
-    get_github_build_metrics,
-    get_table,
-    includes_common_docs,
     is_citable,
-    measure_coverage,
 )
+from almanack.metrics.garden_lattice.practicality import (
+    count_repo_tags,
+    get_ecosystems_package_metrics,
+)
+from almanack.metrics.garden_lattice.understanding import includes_common_docs
 from tests.data.almanack.repo_setup.create_repo import repo_setup
 
 DATETIME_NOW = datetime.now()
